# Reduce worst-case alg complexity of MemoryExtensions.IndexOfAny