Based on a comment thread at https://github.com/dotnet/runtime/pull/53115#discussion_r638016962.

In a nutshell, given a large enough needle, the code pattern below may have `O(n^2 * l)` complexity, where _n_ is the length of the haystack and _l_ is the number of needles.

```cs
Span<T> haystack = GetHaystack();
Span<T> needles = GetNeedles();

while (true)
{
    int idxOfFirstNeedle = haystack.IndexOfAny(needles);
    if (idxOfFirstNeedle >= 0)
    {
        DoSomethingWith(haystack.Slice(0, idxOfFirstNeedle);
        haystack = haystack.Slice(idxOfFirstNeedle + 1);
    }
    else
    {
        DoSomethingWith(haystack); // remainder
        break;
    }
}
```

This PR reduces the algorithmic complexity of this slice-and-loop pattern to worst-case `O(n * l)`, which is what most callers probably expect. The downside is that there's a higher constant factor, and this higher constant factor will dominate when the haystack is small or when first needle has a much higher probability of being found than any subsequent needle.

I'm intentionally keeping this PR simple, and a future PR can introduce vectorization or any other optimization technique to help get some of this perf back. The unit tests introduced as part of this PR should help catch future regressions in this area.

I don't have perf numbers right now but should be able to share them by the end of the week if there's interest in seeing them. Per investigation in the previously linked thread, I don't expect this to impact typical libraries code, as `MemoryExtensions.[Last]IndexOfAny` already has logic to handle specific needle counts typical of perf-critical code paths.